rag-support-bot/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sentence-transformers
faiss-cpu
python-dotenv
